requirements.txt
google-generativeai
python-dotenv
numpy
scipy